+++ example/android/app/src/main/kotlin/com/studioseizh/seizhtv/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.example
+package com.studioseizh.seizhtv
 
 import io.flutter.embedding.android.FlutterActivity